first try

+++ src/ltl-compare/include/ltl-compare.hh
@@ -1,4 +1,4 @@
 
 #include <string>
 //spot includes
-double semanticComparison(std::string a1, std::string a2);
+float  semanticComparison(std::string a1, std::string a2);
